server/requirements.txt
anthropic==0.79.0
fastapi==0.116.1
google-generativeai==0.8.6
langchain==1.2.10
langchain-anthropic==1.3.2
langchain-core==1.2.10
langchain-google-genai==4.2.0
langchain-openai==1.1.8
langgraph==1.0.8
openai==2.19.0
python-dotenv==1.2.1
uvicorn==0.35.0